Fix pg_upgrade for same-major version PostgreSQL upgrades
Adds logic to either rename or move the data directory instead of running
through the whole pg_upgrade process. If we're on the same major version as
the default packaged PostgreSQL release, we try the following:

1) Try to rename the data directory's parent to the proper major version
number. This is fastest, avoids any data moves through a simple filesystem
rename operation, and does not require any additonal space, to boot.

2) If that won't work, we then try to move the /data directory into the
new parent directory. Depending on if the new parent is on another
partition/drive or such, it may require some time to move the files. We
check for available disk space at the destination just in case.

3) Fall back to doing pg_upgrade as a last resort, and perform the upgrade
that way. We should never have to do this if we're on the same major
release.

If the old data directory is on a different PostgreSQL major release, the
usual pg_upgrade path is the only way.

Also:
* Make sure we're not trying to rmdir the old dir unless empty
* Strip out unused old_data_dir comparison already in place with filter
* Add detection to avoid mountpoint renames
* Comment updates

Signed-off-by: Aaron Pavely <aaron@pavely.net>